pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["Daniil <39118675+TyPaporotnyk@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "test_project"}]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.12.0"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.20"
aiogram = "2.25.1"
loguru = "^0.7.1"
redis = "^5.0.0"
psycopg2-binary = "^2.9.7"
asyncpg = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
